+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.5 (SNAPSHOT)
+- improved `get` mojo, added auto-search of CVS folder in package folder hierarchy, it can be disabled with `<disableCvsAutosearch>true</disableCvsAutosearch>`, #23
+- improved `get` mojo, added way to define relative path to CVS folder in `src` folder through `<relativePathToCvsFolder>`, by default the path extracted from package name
+
 # 2.1.4 (24-jun-2017)
 - added support of BAZAAR CVS (experimental)
 - fixed order of processing of CVS branch, tag and revision in `get` mojo